feat: support es3 target
closes #959
